Modulegraph: utils: cleanup compat.
